validation on samples
